requirements.txt
python-dotenv==0.9.1
tweepy==3.7.0
lxml==4.1.1
google-cloud==0.34.0
google-cloud-storage==1.13.2
werkzeug==0.14.1